Make perm parameter on Txn.go obligatory
When passing ?perm:None Txn.go would return a weak type.
This would be ok for now, but only for a default of Rw.
The same scheme would not work for Map.open_existing, because here a
default of Nodup would not type as needed. Therefore make the Dup / Nodup
and Ro / Rw parameters obligatory. Using optional parameters in those
cases would commit us to unordinary phantom types.