reproduce #43
This bug is triggered by lmdb and GC both releasing the same dbi slot.
A finaliser is attached to map handles, which releases the dbi slot.
But when a transaction is aborted, then all dbi slots allocated in that
transaction are released by lmdb. If those slots are reused before the
GC releases them again via the finaliser, then the finaliser releases
dbi slots that are still in use.