further test exhausted max_maps
When the txn is committed after max_maps is exhausted, later on MDB_DBS_FULL will be raised.

But when the txn is aborted after max_maps is exhausted, then later on Invalid_argument "Lmdb" will be raised.
This should not happen, since an aborted txn should have no side-effects.