Do not finalise maps
As Armael pointed out, relying on GC for releasing dbi slots is too fragile.

It turns out that finalising dbi slots without causing breakage in some border cases is
difficult. It is also not necessary.

If max_maps is large enough to hold all maps at once, then there is no
need to release dbi slots.
Otherwise we need to juggle with the available slots and the GC is no
good tool to accomplish this.

fixes #43